requirements.txt
numpy==1.26.0
torch==2.0.1
torchvision==0.15.2
pillow==11.0.0